requirements.txt
tensorflow-gpu==2.1.0
tensorflow-federated==0.13.1